requirements.txt
fastapi==0.115.6
starlette==0.40.0
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.7.1
email-validator==2.2.0
python-dotenv==1.0.1
anyio==4.8.0
psycopg[binary]==3.2.3
redis==5.2.1
loguru==0.7.3
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pytest==8.3.4
httpx==0.27.2
resend==2.0.0
slowapi==0.1.9
